+++ csrc/pytorch/moe_permute/moe_permute.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2026, Advanced Micro Devices, Inc. All rights reserved.
 //
 // See LICENSE for license information.
-#include "primus_turbo/permute.h"
+#include "primus_turbo/moe_permute.h"
 #include "../extensions.h"
 #include "primus_turbo/arch.h"
