+++ primus_turbo/pytorch/ops/moe/__init__.py
@@ -7,6 +7,6 @@
 from .fused_moe_router import *
 from .indices_converter import *
 from .moe_dispatch_combine import *
+from .moe_permute import *
 from .permutation import *
-from .permute import *
 from .tokens_per_expert_to_mask import *
